♻️Refactor: Improved Router structure and organization.
