+++ backend/src/PeopleHub.Application/UseCases/Individual/DeleteIndividualUseCase.cs
@@ -1,50 +1,51 @@
 ﻿using Microsoft.AspNetCore.Http;
 using PeopleHub.Application.Actions;
-using PeopleHub.Application.Dtos.LegalPerson;
+using PeopleHub.Application.Dtos.IndividualPerson;
 using PeopleHub.Application.Dtos.Response;
 using PeopleHub.Application.Interfaces.Common;
-using PeopleHub.Application.Interfaces.LegalPerson;
 using PeopleHub.Application.Interfaces.Log;
 using PeopleHub.Application.Interfaces.UserAccount;
 using PeopleHub.Application.UseCases.Base;
+using PeopleHub.Application.UseCases.Individual.Interfaces;
 using PeopleHub.Domain.Interfaces;
 
-namespace PeopleHub.Application.UseCases.LegalPerson;
+namespace PeopleHub.Application.UseCases.Individual;
 
-public class DeleteLegalPersonUseCase : BaseLoggingUseCase, IDeleteLegalPersonUseCase
+public class DeleteIndividualUseCase : BaseLoggingUseCase, IDeleteIndividualUseCase
 {
     private readonly IPersonRepository _personRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-    public DeleteLegalPersonUseCase(
+
+    public DeleteIndividualUseCase(
         IPersonRepository personRepository, 
         IUnitOfWork unitOfWork, 
-        IAuditLogService auditLogService, 
+        IAuthenticatedUserAccountService authenticatedUserService, 
         IHttpContextAccessor httpContextAccessor, 
-        IAuthenticatedUserAccountService authenticatedUserService,
+        IAuditLogService auditLogService,
         IContextProvider contextProvider) : base(auditLogService, httpContextAccessor, authenticatedUserService, contextProvider)
     {
         _personRepository = personRepository;
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<ApiResponseDto<bool>> ExecuteAsync(DeleteLegalPersonDto request)
+    public async Task<ApiResponseDto<bool>> ExecuteAsync(DeleteIndividualPersonDto request)
     {
         try
         {
-            var person = await _personRepository.GetLegalByCnpjAsync(request.Cnpj);
+            var person = await _personRepository.GetIndividualByCpfAsync(request.Cpf);
             if (person == null)
                 return await ResponseAsync<bool>(
                     logAction: LogAction.NOT_FOUND,
                     eventValue: request,
-                    message: "Legal Person not found."
+                    message: "Individual Person not found."
                 );
 
-            await _personRepository.DeleteLegalAsync(person);
+            await _personRepository.DeleteIndividualAsync(person);
             await _unitOfWork.CommitAsync();
 
             return await ResponseAsync<bool>(
-                logAction: Actions.LogAction.DELETE,
+                logAction: LogAction.DELETE,
                 eventValue: request,
                 oldValue: person,
                 message: "Legal Person has been successfully removed."
