+++ backend/src/PeopleHub.Application/UseCases/Legal/Interfaces/IRegisterLegalUseCase.cs
@@ -1,9 +1,9 @@
 ﻿using PeopleHub.Application.Dtos.LegalPerson;
 using PeopleHub.Application.Dtos.Response;
 
-namespace PeopleHub.Application.Interfaces.LegalPerson;
+namespace PeopleHub.Application.UseCases.Legal.Interfaces;
 
-public interface IRegisterLegalPersonUseCase
+public interface IRegisterLegalUseCase
 {
     Task<ApiResponseDto<bool>> ExecuteAsync(RegisterLegalPersonRequestDto request);
 }
