+++ backend/src/PeopleHub.Application/UseCases/Legal/Interfaces/IUpdateLegalUseCase.cs
@@ -1,9 +1,9 @@
 ﻿using PeopleHub.Application.Dtos.LegalPerson;
 using PeopleHub.Application.Dtos.Response;
 
-namespace PeopleHub.Application.Interfaces.LegalPerson;
+namespace PeopleHub.Application.UseCases.Legal.Interfaces;
 
-public interface IUpdateLegalPersonUseCase
+public interface IUpdateLegalUseCase
 {
     Task<ApiResponseDto<bool>> ExecuteAsync(UpdateLegalPersonRequestDto request);
 }
