+++ backend/src/PeopleHub.Application/UseCases/Legal/RegisterLegalUseCase.cs
@@ -1,28 +1,28 @@
 ﻿using Microsoft.AspNetCore.Http;
 using PeopleHub.Application.Actions;
-using PeopleHub.Application.Dtos.IndividualPerson;
+using PeopleHub.Application.Dtos.LegalPerson;
 using PeopleHub.Application.Dtos.Response;
 using PeopleHub.Application.Interfaces.Common;
-using PeopleHub.Application.Interfaces.IndividualPerson;
 using PeopleHub.Application.Interfaces.Log;
 using PeopleHub.Application.Interfaces.UserAccount;
 using PeopleHub.Application.UseCases.Base;
+using PeopleHub.Application.UseCases.Legal.Interfaces;
 using PeopleHub.Domain.Entities;
 using PeopleHub.Domain.Interfaces;
 using PeopleHub.Domain.ValueObjects;
 
-namespace PeopleHub.Application.UseCases.IndividualPerson;
+namespace PeopleHub.Application.UseCases.Legal;
 
-public class RegisterIndividualPersonUseCase : BaseLoggingUseCase, IRegisterIndividualPersonUseCase
+public class RegisterLegalUseCase : BaseLoggingUseCase, IRegisterLegalUseCase
 {
     private readonly IPersonRepository _personRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-    public RegisterIndividualPersonUseCase(
-        IPersonRepository personRepository,
-        IUnitOfWork unitOfWork,
-        IAuditLogService auditLogService,
-        IHttpContextAccessor httpContextAccessor,
+    public RegisterLegalUseCase(
+        IPersonRepository personRepository, 
+        IUnitOfWork unitOfWork, 
+        IAuditLogService auditLogService, 
+        IHttpContextAccessor httpContextAccessor, 
         IAuthenticatedUserAccountService authenticatedUserService,
         IContextProvider contextProvider) : base(auditLogService, httpContextAccessor, authenticatedUserService, contextProvider)
     {
@@ -30,11 +30,11 @@ public class RegisterIndividualPersonUseCase : BaseLoggingUseCase, IRegisterIndi
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<ApiResponseDto<bool>> ExecuteAsync(RegisterIndividualPersonRequestDto request)
-    { 
+    public async Task<ApiResponseDto<bool>> ExecuteAsync(RegisterLegalPersonRequestDto request)
+    {
         try
         {
-            var existingPerson = await _personRepository.GetIndividualByCpfAsync(request.Cpf);
+            var existingPerson = await _personRepository.GetLegalByCnpjAsync(request.Cnpj);
 
             if (existingPerson != null)
                 return await ResponseAsync<bool>(
@@ -48,17 +48,20 @@ public class RegisterIndividualPersonUseCase : BaseLoggingUseCase, IRegisterIndi
             var address = new Address(request.Street, request.Number, request.Complement, request.City, request.State, request.ZipCode);
             var phone = new Phone(request.Phone);
             var email = new Email(request.Email);
-            var cpf = new Cpf(request.Cpf);
+            var cnpj = new Cnpj(request.Cnpj);
+            var cpf = new Cpf(request.LegalRepresentativeCpf);
 
-            var person = new IndividualPersonEntity(
-                request.FullName,
-                cpf,
-                request.BirthDate,
-                request.Gender,
+            var person = new LegalPersonEntity(
+                request.LegalName,
+                request.TradeName,
+                cnpj,
+                request.StateRegistration,
+                request.MunicipalRegistration,
                 address,
                 phone,
-                email
-            );
+                email,
+                request.LegalRepresentativeName,
+                cpf);
 
             await _personRepository.AddAsync(person);
             await _unitOfWork.CommitAsync();
@@ -66,12 +69,13 @@ public class RegisterIndividualPersonUseCase : BaseLoggingUseCase, IRegisterIndi
             return await ResponseAsync<bool>(
                 logAction: LogAction.CREATE,
                 eventValue: person,
-                message: "Individual Person successfully registered."
+                message: "Legal Person successfully registered."
             );
         }
         catch (Exception ex)
         {
             return await ResponseAsync<bool>(logAction: LogAction.ERROR, message: ex.Message);
         }
+
     }
 }
