+++ backend/src/PeopleHub.Application/UseCases/Legal/UpdateLegalUseCase.cs
@@ -1,26 +1,26 @@
 ﻿using Microsoft.AspNetCore.Http;
 using PeopleHub.Application.Actions;
-using PeopleHub.Application.Dtos.IndividualPerson;
+using PeopleHub.Application.Dtos.LegalPerson;
 using PeopleHub.Application.Dtos.Response;
 using PeopleHub.Application.Interfaces.Common;
-using PeopleHub.Application.Interfaces.IndividualPerson;
 using PeopleHub.Application.Interfaces.Log;
 using PeopleHub.Application.Interfaces.UserAccount;
 using PeopleHub.Application.UseCases.Base;
+using PeopleHub.Application.UseCases.Legal.Interfaces;
 using PeopleHub.Domain.Interfaces;
 using PeopleHub.Domain.ValueObjects;
 
-namespace PeopleHub.Application.UseCases.IndividualPerson;
+namespace PeopleHub.Application.UseCases.Legal;
 
-public class UpdateIndividualPersonUseCase : BaseLoggingUseCase, IUpdateIndividualPersonUseCase
+public class UpdateLegalUseCase : BaseLoggingUseCase, IUpdateLegalUseCase
 {
     private readonly IPersonRepository _personRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-    public UpdateIndividualPersonUseCase(
+    public UpdateLegalUseCase(
         IPersonRepository personRepository, 
         IUnitOfWork unitOfWork, 
-        IAuditLogService auditLogService, 
+        IAuditLogService auditLogService,
         IHttpContextAccessor httpContextAccessor, 
         IAuthenticatedUserAccountService authenticatedUserService,
         IContextProvider contextProvider) : base(auditLogService, httpContextAccessor, authenticatedUserService, contextProvider)
@@ -29,40 +29,38 @@ public class UpdateIndividualPersonUseCase : BaseLoggingUseCase, IUpdateIndividu
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<ApiResponseDto<bool>> ExecuteAsync(UpdateIndividualPersonRequestDto request)
+    public async Task<ApiResponseDto<bool>> ExecuteAsync(UpdateLegalPersonRequestDto request)
     {
         try
         {
-            var person = await _personRepository.GetIndividualByCpfAsync(request.Cpf);
+            var person = await _personRepository.GetLegalByCnpjAsync(request.Cnpj);
             if (person == null)
                 return await ResponseAsync<bool>(
                     logAction: LogAction.NOT_FOUND,
                     eventValue: request,
-                    message: "Individual Person not found."
+                    message: "Legal Person not found."
                 );
 
-            var address = new Address(request.Street, request.Number, request.Complement, request.City, request.State, request.ZipCode);
-            var phone = new Phone(request.Phone);
-            var email = new Email(request.Email);
-            var cpf = new Cpf(request.Cpf);
-
-            person.UpdateIndividualPerson(
-                request.FullName,
-                request.BirthDate,
-                request.Gender,
-                address,
-                phone,
-                email
+            person.UpdateLegalPerson(
+                request.LegalName,
+                request.TradeName,
+                request.StateRegistration,
+                request.MunicipalRegistration,
+                new Address(request.Street, request.Number, request.Complement, request.City, request.State, request.ZipCode),
+                new Phone(request.Phone),
+                new Email(request.Email),
+                request.LegalRepresentativeName,
+                new Cpf(request.LegalRepresentativeCpf)
             );
 
-            await _personRepository.UpdateIndividualAsync(person);
+            await _personRepository.UpdateLegalAsync(person);
             await _unitOfWork.CommitAsync();
 
             return await ResponseAsync<bool>(
                 logAction: LogAction.UPDATE,
                 eventValue: request,
                 oldValue: person,
-                message: "Individual Person updated successfully."
+                message: "Legal Person updated successfully."
             );
         }
         catch (Exception ex)
